// kraken-web/src/gen/kraken_pb.ts
export enum TaskStatus {
  /**
   * @generated from enum value: TASK_STATUS_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * Waiting for implant check-in
   *
   * @generated from enum value: TASK_STATUS_QUEUED = 1;
   */
  QUEUED = 1,

  /**
   * Sent to implant, awaiting result
   *
   * @generated from enum value: TASK_STATUS_DISPATCHED = 2;
   */
  DISPATCHED = 2,

  /**
   * Success
   *
   * @generated from enum value: TASK_STATUS_COMPLETED = 3;
   */
  COMPLETED = 3,

  /**
   * Error
   *
   * @generated from enum value: TASK_STATUS_FAILED = 4;
   */
  FAILED = 4,

  /**
   * Operator cancelled
   *
   * @generated from enum value: TASK_STATUS_CANCELLED = 5;
   */
  CANCELLED = 5,

  /**
   * Implant lost or timeout
   *
   * @generated from enum value: TASK_STATUS_EXPIRED = 6;
   */
  EXPIRED = 6,
}
// ...
export class FileTask extends Message<FileTask> {
  /**
   * @generated from oneof kraken.FileTask.operation
   */
  operation: {
    /**
     * @generated from field: kraken.FileList list = 1;
     */
    value: FileList;
    case: "list";
  } | {
    /**
     * @generated from field: kraken.FileRead read = 2;
     */
    value: FileRead;
    case: "read";
  } | {
    /**
     * @generated from field: kraken.FileWrite write = 3;
     */
    value: FileWrite;
    case: "write";
  } | {
    /**
     * @generated from field: kraken.FileDelete delete = 4;
     */
    value: FileDelete;
    case: "delete";
  } | {
    /**
     * @generated from field: kraken.FileUpload upload = 5;
     */
    value: FileUpload;
    case: "upload";
  } | {
    /**
     * @generated from field: kraken.FileDownload download = 6;
     */
    value: FileDownload;
    case: "download";
  } | {
    /**
     * @generated from field: kraken.FileUploadChunked upload_chunked = 7;
     */
    value: FileUploadChunked;
    case: "uploadChunked";
  } | {
    /**
     * @generated from field: kraken.FileDownloadChunked download_chunked = 8;
     */
    value: FileDownloadChunked;
    case: "downloadChunked";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<FileTask>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.FileTask";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "list", kind: "message", T: FileList, oneof: "operation" },
    { no: 2, name: "read", kind: "message", T: FileRead, oneof: "operation" },
    { no: 3, name: "write", kind: "message", T: FileWrite, oneof: "operation" },
    { no: 4, name: "delete", kind: "message", T: FileDelete, oneof: "operation" },
    { no: 5, name: "upload", kind: "message", T: FileUpload, oneof: "operation" },
    { no: 6, name: "download", kind: "message", T: FileDownload, oneof: "operation" },
    { no: 7, name: "upload_chunked", kind: "message", T: FileUploadChunked, oneof: "operation" },
    { no: 8, name: "download_chunked", kind: "message", T: FileDownloadChunked, oneof: "operation" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileTask {
    return new FileTask().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileTask {
    return new FileTask().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileTask {
    return new FileTask().fromJsonString(jsonString, options);
  }

  static equals(a: FileTask | PlainMessage<FileTask> | undefined, b: FileTask | PlainMessage<FileTask> | undefined): boolean {
    return proto3.util.equals(FileTask, a, b);
  }
}

/**
 * @generated from message kraken.FileList
 */
export class FileList extends Message<FileList> {
  /**
   * @generated from field: string path = 1;
   */
  path = "";

  /**
   * @generated from field: bool recursive = 2;
   */
  recursive = false;

  constructor(data?: PartialMessage<FileList>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.FileList";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "recursive", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileList {
    return new FileList().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileList {
    return new FileList().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileList {
    return new FileList().fromJsonString(jsonString, options);
  }

  static equals(a: FileList | PlainMessage<FileList> | undefined, b: FileList | PlainMessage<FileList> | undefined): boolean {
    return proto3.util.equals(FileList, a, b);
  }
}

/**
 * @generated from message kraken.FileRead
 */
export class FileRead extends Message<FileRead> {
  /**
   * @generated from field: string path = 1;
   */
  path = "";

  /**
   * @generated from field: optional uint64 offset = 2;
   */
  offset?: bigint;

  /**
   * @generated from field: optional uint64 length = 3;
   */
  length?: bigint;

  constructor(data?: PartialMessage<FileRead>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.FileRead";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "offset", kind: "scalar", T: 4 /* ScalarType.UINT64 */, opt: true },
    { no: 3, name: "length", kind: "scalar", T: 4 /* ScalarType.UINT64 */, opt: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileRead {
    return new FileRead().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileRead {
    return new FileRead().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileRead {
    return new FileRead().fromJsonString(jsonString, options);
  }

  static equals(a: FileRead | PlainMessage<FileRead> | undefined, b: FileRead | PlainMessage<FileRead> | undefined): boolean {
    return proto3.util.equals(FileRead, a, b);
  }
}
// ...
export class FileUpload extends Message<FileUpload> {
  /**
   * @generated from field: string remote_path = 1;
   */
  remotePath = "";

  /**
   * @generated from field: bytes data = 2;
   */
  data = new Uint8Array(0);

  constructor(data?: PartialMessage<FileUpload>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.FileUpload";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "remote_path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "data", kind: "scalar", T: 12 /* ScalarType.BYTES */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileUpload {
    return new FileUpload().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileUpload {
    return new FileUpload().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileUpload {
    return new FileUpload().fromJsonString(jsonString, options);
  }

  static equals(a: FileUpload | PlainMessage<FileUpload> | undefined, b: FileUpload | PlainMessage<FileUpload> | undefined): boolean {
    return proto3.util.equals(FileUpload, a, b);
  }
}

/**
 * @generated from message kraken.FileDownload
 */
export class FileDownload extends Message<FileDownload> {
  /**
   * @generated from field: string remote_path = 1;
   */
  remotePath = "";

  constructor(data?: PartialMessage<FileDownload>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.FileDownload";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "remote_path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileDownload {
    return new FileDownload().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileDownload {
    return new FileDownload().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileDownload {
    return new FileDownload().fromJsonString(jsonString, options);
  }

  static equals(a: FileDownload | PlainMessage<FileDownload> | undefined, b: FileDownload | PlainMessage<FileDownload> | undefined): boolean {
    return proto3.util.equals(FileDownload, a, b);
  }
}
// ...
export class DirectoryListing extends Message<DirectoryListing> {
  /**
   * @generated from field: string path = 1;
   */
  path = "";

  /**
   * @generated from field: repeated kraken.DirectoryEntry entries = 2;
   */
  entries: DirectoryEntry[] = [];

  constructor(data?: PartialMessage<DirectoryListing>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.DirectoryListing";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "entries", kind: "message", T: DirectoryEntry, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): DirectoryListing {
    return new DirectoryListing().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): DirectoryListing {
    return new DirectoryListing().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): DirectoryListing {
    return new DirectoryListing().fromJsonString(jsonString, options);
  }

  static equals(a: DirectoryListing | PlainMessage<DirectoryListing> | undefined, b: DirectoryListing | PlainMessage<DirectoryListing> | undefined): boolean {
    return proto3.util.equals(DirectoryListing, a, b);
  }
}
// ...
export class FileContents extends Message<FileContents> {
  /**
   * @generated from field: string path = 1;
   */
  path = "";

  /**
   * @generated from field: bytes data = 2;
   */
  data = new Uint8Array(0);

  /**
   * @generated from field: uint64 size = 3;
   */
  size = protoInt64.zero;

  constructor(data?: PartialMessage<FileContents>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "kraken.FileContents";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "data", kind: "scalar", T: 12 /* ScalarType.BYTES */ },
    { no: 3, name: "size", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileContents {
    return new FileContents().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileContents {
    return new FileContents().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileContents {
    return new FileContents().fromJsonString(jsonString, options);
  }

  static equals(a: FileContents | PlainMessage<FileContents> | undefined, b: FileContents | PlainMessage<FileContents> | undefined): boolean {
    return proto3.util.equals(FileContents, a, b);
  }
}